separate type checking logic
