+++ ksqldb-functional-tests/src/test/resources/query-validation-tests/group-by.json
@@ -602,7 +602,7 @@
       "name": "multiple expressions - delimited key format to multi-column primitive format",
       "statements": [
         "CREATE STREAM TEST (f1 STRING KEY, f2 INT) WITH (kafka_topic='test_topic', key_format='DELIMITED', value_format='JSON', partitions=1);",
-        "CREATE TABLE OUTPUT WITH (KEY_FORMAT='JSON') AS SELECT f1 AS key, COUNT(*) AS total FROM TEST GROUP BY f1;"
+        "CREATE TABLE OUTPUT AS SELECT f1 AS k1, f2 AS k2, COUNT(*) AS total FROM TEST GROUP BY f1, f2;"
       ],
       "inputs": [
         {"topic": "test_topic", "key": "a", "value": {"f2": "2"}},
@@ -615,12 +615,36 @@
         {"topic": "test_topic", "key": "b", "value": {"f2": "1"}}
       ],
       "outputs": [
-        {"topic": "OUTPUT", "key": "a", "value": {"TOTAL": 1}},
-        {"topic": "OUTPUT", "key": "b", "value": {"TOTAL": 1}},
-        {"topic": "OUTPUT", "key": "a", "value": {"TOTAL": 2}},
-        {"topic": "OUTPUT", "key": "a", "value": {"TOTAL": 3}},
-        {"topic": "OUTPUT", "key": "b", "value": {"TOTAL": 2}},
-        {"topic": "OUTPUT", "key": "b", "value": {"TOTAL": 3}}
+        {"topic": "OUTPUT", "key":"a,2", "value": {"TOTAL": 1}},
+        {"topic": "OUTPUT", "key":"b,4", "value": {"TOTAL": 1}},
+        {"topic": "OUTPUT", "key":"a,2", "value": {"TOTAL": 2}},
+        {"topic": "OUTPUT", "key":"b,4", "value": {"TOTAL": 2}},
+        {"topic": "OUTPUT", "key":"b,1", "value": {"TOTAL": 1}}
+      ]
+    },
+    {
+      "name": "multiple expressions - delimited key format to multi-column, non-primitive format",
+      "statements": [
+        "CREATE STREAM TEST (f1 STRING KEY, f2 INT) WITH (kafka_topic='test_topic', key_format='DELIMITED', value_format='JSON', partitions=1);",
+        "CREATE TABLE OUTPUT WITH (KEY_FORMAT='JSON') AS SELECT STRUCT(f1:=f1, f2:=f2) AS k1, f1 as k2, as_value(struct(f1:=f1, f2:=f2)) AS key, COUNT(*) AS total FROM TEST GROUP BY STRUCT(f1:=f1, f2:=f2), f1;"
+      ],
+      "inputs": [
+        {"topic": "test_topic", "key": "a", "value": {"f2": "2"}},
+        {"topic": "test_topic", "key": "b", "value": {"f2": "4"}},
+        {"topic": "test_topic", "key": "a", "value": {"f2": null}},
+        {"topic": "test_topic", "key": "a", "value": null},
+        {"topic": "test_topic", "key": null, "value": {"f2": "1"}},
+        {"topic": "test_topic", "key": "a", "value": {"f2": "2"}},
+        {"topic": "test_topic", "key": "b", "value": {"f2": "4"}},
+        {"topic": "test_topic", "key": "b", "value": {"f2": "1"}}
+      ],
+      "outputs": [
+        {"topic": "OUTPUT", "key": {"K1": {"F1": "a", "F2": 2}, "K2": "a"}, "value": {"KEY": {"F1":"a","F2":2}, "TOTAL": 1}},
+        {"topic": "OUTPUT", "key": {"K1": {"F1": "b", "F2": 4}, "K2": "b"}, "value": {"KEY": {"F1":"b","F2":4}, "TOTAL": 1}},
+        {"topic": "OUTPUT", "key": {"K1": {"F1": "a", "F2": null}, "K2": "a"}, "value": {"KEY": {"F1":"a","F2":null}, "TOTAL": 1}},
+        {"topic": "OUTPUT", "key": {"K1": {"F1": "a", "F2": 2}, "K2": "a"}, "value": {"KEY": {"F1":"a","F2":2}, "TOTAL": 2}},
+        {"topic": "OUTPUT", "key": {"K1": {"F1": "b", "F2": 4}, "K2": "b"}, "value": {"KEY": {"F1":"b","F2":4}, "TOTAL": 2}},
+        {"topic": "OUTPUT", "key": {"K1": {"F1": "b", "F2": 1}, "K2": "b"}, "value": {"KEY": {"F1":"b","F2":1}, "TOTAL": 1}}
       ]
     },
     {
